Extract app to separate modules, add dependency management
